+++ app/src/main/java/net/weero/measix/pilot/ui/components/message/tools/ToolUI.kt
@@ -100,6 +100,8 @@ object ToolUIRegistry {
         RecentChatsToolUI,
         ConversationSearchToolUI,
         GetScreenTimeToolUI,
+        CalendarQueryToolUI,
+        CalendarCreateToolUI,
         EditFileToolUI,
         ReadFileToolUI,
         WriteFileToolUI,